requirements.txt
paho-mqtt==1.2
wheel==0.24.0
